+++ core/services/ocr2/plugins/ccip/internal/ccipdata/offramp_v1_0_0.go
@@ -24,6 +24,7 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/abihelpers"
 	ccipconfig "github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/config"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal"
+	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/collections"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/logpollerutil"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/rpclib"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/prices"
@@ -176,7 +177,7 @@ func (o *OffRampV1_0_0) GetDestinationTokensFromSourceTokens(ctx context.Context
 		return nil, fmt.Errorf("parse outputs: %w", err)
 	}
 
-	seenDestTokens := internal.NewSet[common.Address]()
+	seenDestTokens := collections.NewSet[common.Address]()
 	for _, destToken := range destTokens {
 		if seenDestTokens.Contains(destToken) {
 			return nil, fmt.Errorf("offRamp misconfig, destination token %s already exists", destToken)
