+++ core/services/ocr2/plugins/ccip/internal/collections/set.go
@@ -1,4 +1,4 @@
-package internal
+package collections
 
 import "sync"
